requirements.txt
discord
humanfriendly
google_cloud_secret_manager
flask